Move lastError into utils.go
