+++ components/movementsensor/utils.go
@@ -3,6 +3,7 @@ package movementsensor
 import (
 	"errors"
 	"math"
+	"sync"
 
 	geo "github.com/kellydunn/golang-geo"
 
@@ -69,3 +70,24 @@ var (
 	// ErrMethodUnimplementedLinearAcceleration returns error if Linear Acceleration is unimplemented.
 	ErrMethodUnimplementedLinearAcceleration = errors.New("linear acceleration unimplemented")
 )
+
+type lastError struct {
+	err error
+	mu  sync.Mutex
+}
+
+func (le *lastError) Set(err error) {
+	le.mu.Lock()
+	defer le.mu.Unlock()
+	le.err = err
+}
+
+func (le *lastError) Get() error {
+	le.mu.Lock()
+	defer le.mu.Unlock()
+
+	err := le.err
+	le.err = nil
+	return err
+}
+
